pyproject.toml
[tool.poetry]
name = "books"
version = "0.1.0"
description = ""
authors = ["Alina <fakherdinova.alya@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg2 = "^2.9.10"
pydantic-config = "^0.3.0"
django-crispy-forms = "^2.3"
crispy-bootstrap4 = "^2024.10"
django-contrib-postgres = "^0.0.1"
pillow = "^11.0.0"
djangoajax = "^3.3"
django-ajax-selects = "^3.0.3"
django-select2 = "^8.2.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
